requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.11.0
h11==0.16.0
idna==3.7
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
pycparser==2.23
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv==1.2.1
requests==2.32.3
ruff==0.14.3
selenium==4.38.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.2.0
